Implemented Chess\SanTrait
