+++ src/SanHeuristics.php
@@ -2,8 +2,6 @@
 
 namespace Chess;
 
-use Chess\Eval\AbstractEval;
-use Chess\Eval\InverseEvalInterface;
 use Chess\Function\AbstractFunction;
 use Chess\Play\SanPlay;
 use Chess\Variant\AbstractBoard;
@@ -17,6 +15,8 @@ use Chess\Variant\Classical\PGN\AN\Color;
  */
 class SanHeuristics extends SanPlay
 {
+    use SanTrait;
+
     /**
      * Function.
      *
@@ -62,7 +62,9 @@ class SanHeuristics extends SanPlay
         $this->function = $function;
         $this->name = $name;
 
-        $this->calc()->balance()->normalize(-1, 1);
+        $this->calc()->balance();
+
+        $this->balance = $this->normalize(-1, 1, $this->balance);
     }
 
     /**
@@ -95,73 +97,11 @@ class SanHeuristics extends SanPlay
 
     /**
      * Balances the result.
-     *
-     * @return \Chess\SanHeuristics
      */
-    protected function balance(): SanHeuristics
+    protected function balance(): void
     {
         foreach ($this->result as $result) {
             $this->balance[] = $result[Color::W] - $result[Color::B];
         }
-
-        return $this;
-    }
-
-    /**
-     * Normalizes the balance.
-     *
-     * @param int $newMin
-     * @param int $newMax
-     */
-    protected function normalize(int $newMin, int $newMax): void
-    {
-        $min = min($this->balance);
-        $max = max($this->balance);
-
-        foreach ($this->balance as $key => $val) {
-            if ($val > 0) {
-                $this->balance[$key] = round($this->balance[$key] * $newMax / $max, 2);
-            } elseif ($val < 0) {
-                $this->balance[$key] = round($this->balance[$key] * $newMin / $min, 2);
-            } else {
-                $this->balance[$key] = 0;
-            }
-        }
-    }
-
-    /**
-     * Calculates an item.
-     *
-     * @param \Chess\Eval\AbstractEval $eval
-     * @return array
-     */
-    protected function item(AbstractEval $eval): array
-    {
-        $result = $eval->result;
-
-        if (is_array($result[Color::W])) {
-            if ($eval instanceof InverseEvalInterface) {
-                $item = [
-                    Color::W => count($result[Color::B]),
-                    Color::B => count($result[Color::W]),
-                ];
-            } else {
-                $item = [
-                    Color::W => count($result[Color::W]),
-                    Color::B => count($result[Color::B]),
-                ];
-            }
-        } else {
-            if ($eval instanceof InverseEvalInterface) {
-                $item = [
-                    Color::W => $result[Color::B],
-                    Color::B => $result[Color::W],
-                ];
-            } else {
-                $item = $result;
-            }
-        }
-
-        return $item;
     }
 }
